requirements.txt
langchain==0.0.331
openai==0.28.1
pandas
numpy
load_dotenv
uvicorn
fastapi
chromadb==0.4.12
tensorflow
transformers
tf-keras